refactor: pearson corr matrix function
